修改 sh (file_result_set) 和socket 初始化的位置，goto 语句告警

+++ src/dbi/ntgs_mysql.cpp
@@ -68,7 +68,7 @@ page_t * get_page(Connection_T con, int index) {
 	if (!ResultSet_next(rst))
 		goto e_result_set;
 	page->url = paser_url(ResultSet_getString(rst, 1),
-			ResultSet_getColumnSize(rst, 1));
+			(int)ResultSet_getColumnSize(rst, 1));
     
 	page->files_result_set = Connection_executeQuery(con,
 			"select file_path from resource_t where page_id = '%d'", index);
